# api/v1/swagger.json
{
	"components": {
...
		"schemas": {
			"Asset": {
				"properties": {
					"id": {
						"type": "string"
					},
					"rank": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"supply": {
						"type": "string",
						"nullable": true
					},
					"maxSupply": {
						"type": "string",
						"nullable": true
					},
					"marketCapUsd": {
						"type": "string",
						"nullable": true
					},
					"volumeUsd24Hr": {
						"type": "string",
						"nullable": true
					},
					"priceUsd": {
						"type": "string",
						"nullable": true
					},
					"changePercent24Hr": {
						"type": "string",
						"nullable": true
					},
					"vwap24Hr": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"rank",
					"symbol",
					"name",
					"supply",
					"maxSupply",
					"marketCapUsd",
					"volumeUsd24Hr",
					"priceUsd",
					"changePercent24Hr",
					"vwap24Hr"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "coincap-wrapper",
		"version": "1.0.0",
		"description": "coincap wrapper",
		"license": {
			"name": "AGPL-3.0-or-later"
		},
		"contact": {
			"name": "Reinder van der Voort"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Asset": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Asset"
									},
									"type": "array"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Entity"
					},
					"500": {
						"description": "Server Error"
					}
				},
				"tags": [
					"Asset"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ids",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"id",
								"rank",
								"symbol",
								"name",
								"supply",
								"maxSupply",
								"marketCapUsd",
								"volumeUsd24Hr",
								"priceUsd",
								"changePercent24Hr",
								"vwap24Hr"
							]
						}
					}
				]
			}
		},
		"/Asset/{id}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Asset"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Entity"
					},
					"500": {
						"description": "Server Error"
					}
				},
				"tags": [
					"Asset"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/Asset/{id}/USD": {
			"get": {
				"operationId": "GetUSD",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable Entity"
					},
					"500": {
						"description": "Server Error"
					}
				},
				"tags": [
					"Asset"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost/api/v1",
			"description": "Local development"
		}
	],
...
}