# src/data-layer/models/asset.ts
export interface Asset {
  id: string;
  rank: string;
  symbol: string;
  name: null | string;
  supply: null | string;
  maxSupply: null | string;
  marketCapUsd: null | string;
  volumeUsd24Hr: null | string;
  priceUsd: null | string;
  changePercent24Hr: null | string;
  vwap24Hr: null | string;
}